alpaca-trading-bot/requirements.txt
# Alpaca Trading
alpaca-py>=0.13.0

# Perplexity API
requests>=2.31.0
aiohttp>=3.9.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
pandas-ta>=0.3.14b0

# Web Frameworks & Async
fastapi>=0.104.0
uvicorn>=0.24.0
websockets>=11.0

# Environment & Configuration
python-dotenv>=1.0.0
pydantic>=2.4.0
pydantic-settings>=2.0.0

# Logging & Monitoring
loguru>=0.7.2
prometheus-client>=0.18.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0

# Development Tools
black>=23.10.0
flake8>=6.1.0
mypy>=1.6.0
pre-commit>=3.5.0